requirements.txt
pillow>=10.0.0
openai>=1.0.0
opencv-python>=4.8.0
PyYAML>=6.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0